#60 feat: 테스트 결과 페이지 로딩 및 data fetching 로직 수정
